refactor(wwa-message): wwa_message をさらに細かく分割 (再)
オリジナル: 22dc7db
